feat: global Bilibili login with cross-server credential sync
- Make Bilibili login/logout global (single account per server/network) using /display login/logout, OP-only
- Broadcast global credential to all online v2 clients on login/logout/refresh
- snapshotFor() now prefers the global credential, falling back to per-player
- Store encrypted credentials (SESSDATA + refresh token) in the same database as displays
  (SQLite locally, or shared MySQL for cross-server sync)
- Add SqlCredentialSyncBackend + CredentialSyncBackend interface
- Add dreamdisplayx.login / dreamdisplayx.logout permission nodes (LuckPerms supported, default OP)
- Remove screenshare (screen_share) config section and CastManager usage
- Fix SettingsSection scrollbar color type for 1.21.11+ (Int vs Long)
- Keep /display login and /display logout commands (now OP-only global)

+++ platform/client/common/src/main/kotlin/com/dreamdisplayx/platform/client/ui/menu/SettingsSection.kt
@@ -168,7 +168,7 @@ class SettingsSection(
         val start = top + (scrollOffset.toFloat() / maxScroll * (viewH - len)).toInt()
         //? if >=1.21.11 {
         val hovered = mouseY >= start && mouseY <= start + len
-        g.fill(barX, start, barX + barW, start + len, if (hovered) 0xFFA0A0A0 else UiTheme.SCROLLBAR_THUMB)
+        g.fill(barX, start, barX + barW, start + len, if (hovered) 0xFFA0A0A0.toInt() else UiTheme.SCROLLBAR_THUMB)
         //?} else
         /*g.fill(barX, start, barX + barW, start + len, UiTheme.SCROLLBAR_THUMB)*/
         thumbStart = start
